% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaglm.R
\name{formulaglm}
\alias{formulaglm}
\title{Devuelve formula en formato >=0 para modelo glm}
\usage{
formulaglm(modelo, spec = 0.8)
}
\arguments{
\item{modelo}{Modelo glm}

\item{spec}{Especificidad}
}
\value{
Formula en formato >=0 para modelo glm
}
\description{
Devuelve formula en formato >=0 para modelo glm
}
