% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fronteraglm.R
\name{fronteraglm}
\alias{fronteraglm}
\title{Grafica frontera de decision para regresion logistica en dos variables}
\usage{
fronteraglm(modelo, data, prob)
}
\arguments{
\item{modelo}{Objeto glm}

\item{data}{Dataframe correspondiente a los puntos a graficar}

\item{prob}{Probabilidad de corte para decidir clase}
}
\value{
Grafico con frontera de decision
}
\description{
Grafica frontera de decision para regresion logistica en dos variables
}
