% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fronteraknn.R
\name{fronteraknn}
\alias{fronteraknn}
\title{Grafica frontera de decision para knn en dos variables}
\usage{
fronteraknn(modelo, data, prob)
}
\arguments{
\item{modelo}{Objeto knn}

\item{data}{Dataframe correspondiente a los puntos a graficar}

\item{prob}{Probabilidad de corte para decidir clase}
}
\value{
Grafico con frontera de decision
}
\description{
Grafica frontera de decision para knn en dos variables
}
