% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fronterasvm.R
\name{fronterasvm}
\alias{fronterasvm}
\title{Grafica frontera de decision para regresion logistica en dos variables}
\usage{
fronterasvm(modelo, data)
}
\arguments{
\item{modelo}{Objeto glm}

\item{data}{Dataframe correspondiente a los puntos a graficar}
}
\value{
Grafico con frontera de decision
}
\description{
Grafica frontera de decision para regresion logistica en dos variables
}
