% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune2.R
\name{prune2}
\alias{prune2}
\title{Poda arbol en profundidad deseada}
\usage{
prune2(arbol, i)
}
\arguments{
\item{arbol}{arbol profundo}

\item{i}{nivel deseado}
}
\description{
Poda arbol en profundidad deseada
}
