% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidacionCruzada.R, R/ValidacionCruzadaglm.R
\name{ValidacionCruzadaglm}
\alias{ValidacionCruzadaglm}
\title{Realiza validacion cruzada de k capas para un modelo glm}
\usage{
ValidacionCruzadaglm(formula, ncapas, specificity, data)

ValidacionCruzadaglm(formula, ncapas, specificity, data)
}
\arguments{
\item{formula}{formula del modelo glm}

\item{ncapas}{Numero de capas a usar}

\item{specificity}{Especificidad (1-falsos positivos) a fijar en el set de entrenamiento para encontrar probabilidad de corte}

\item{data}{Set de datos}

\item{df}{Set de datos}

\item{formula}{formula del modelo glm}

\item{ncapas}{Numero de capas a usar}

\item{specificity}{Especificidad (1-falsos positivos) a fijar en el set de entrenamiento para encontrar probabilidad de corte}
}
\value{
Lista con sensibilidad y especificidad promedio en conjunto de validacion

Lista con sensibilidad y especificidad promedio en conjunto de validacion
}
\description{
Realiza validacion cruzada de k capas para un modelo glm

Realiza validacion cruzada de k capas para un modelo glm
}
