% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probcorte.R
\name{probcorte}
\alias{probcorte}
\title{Grafica frontera de decision para regresion logistica en dos variables}
\usage{
probcorte(modelo, data, spec)
}
\arguments{
\item{modelo}{Objeto glm}

\item{data}{Dataframe con datos}

\item{spec}{Especificidad buscada}
}
\value{
Curva ROC, AUC y probabilidad de corte
}
\description{
Grafica frontera de decision para regresion logistica en dos variables
}
