% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probcorte.R
\name{probcorte}
\alias{probcorte}
\title{Obtiene probabilidad de corte para especifidad determinada}
\usage{
probcorte(modelo, data, spec, tipo = "glm", grafico = T)
}
\arguments{
\item{modelo}{Objeto glm}

\item{data}{Dataframe con datos}

\item{spec}{Especificidad buscada}
}
\value{
Curva ROC, AUC y probabilidad de corte
}
\description{
Obtiene probabilidad de corte para especifidad determinada
}
